Rename the file for the service checker
