+++ lib/kumo_dockercloud.rb
@@ -2,5 +2,5 @@ require 'kumo_dockercloud/environment'
 require 'kumo_dockercloud/deployment'
 require 'kumo_dockercloud/stack'
 require 'kumo_dockercloud/service'
-require 'kumo_dockercloud/service_check'
+require 'kumo_dockercloud/service_checker'
 require 'kumo_dockercloud/errors'
