+++ lib/kumo_dockercloud/service_checker.rb
