+++ spec/kumo_dockercloud/service_checker_spec.rb
